refactor(goal_curation/operations): split tests to new sibling tests_operations.rs (#1266)
Inline test module in operations.rs is moved into NEW sibling tests_operations.rs.
Existing tests.rs (243 LOC) was not extended to avoid pushing it close to 400.

- src/goal_curation/operations.rs: 480 → 303 LOC (under 400-line rule)
- src/goal_curation/tests_operations.rs: new (~175 LOC)
- src/goal_curation/mod.rs: register mod tests_operations
- super::* → super::operations::* + use super::types::{...}

Pure refactor: zero logic changes, all 13 tests pass.

Closes 12/70 of #1266 (this loop iteration), 21 PRs cumulative since #1285.
